% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{create_muggle}
\alias{create_muggle}
\title{Create a muggle project}
\usage{
create_muggle(
  path,
  fields = list(),
  license = usethis::use_mit_license,
  license_holder = character(),
  organisation = NULL,
  private = FALSE
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}

\item{license}{one of the license functions in \link{usethis}}

\item{license_holder}{giving the license holder, used as \code{cph} and \code{fnd} role in \code{DESCRIPTION}}

\item{organisation}{If supplied, the repo will be created under this
organisation, instead of the account of the user associated with the
\code{auth_token}. You must have permission to create repositories.}

\item{private}{If \code{TRUE}, creates a private repository.}
}
\description{
Set up, or migrate to a muggle project.
Wraps the following steps, \emph{if the respective files or configuration do not already exist}:
\enumerate{
\item Sets up package scaffolding via \code{\link[usethis:create_package]{usethis::create_package()}}.
\item Initialises a git repo via \code{\link[usethis:use_git]{usethis::use_git()}}.
\item Creates a repo on GitHub and sets it as an origin remote.
\item Adds a \code{README.md} via \code{\link[usethis:use_readme_rmd]{usethis::use_readme_md()}}.
\item Sets up the project for unit tests via \code{\link[usethis:use_testthat]{usethis::use_testthat()}} and test coverage via \code{\link[usethis:use_coverage]{usethis::use_coverage()}}.
\item Adds a pkgdown website via \code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}}.
\item Opens the \code{DESCRIPTION} and \code{README.md} for additional edits.
}
}
\section{warning}{
\itemize{
\item Must not be run \emph{inside} a package, but at the root of all packages
\item If run on an existing project, the project should be under version control, with a clean working tree.
The user should check all changes.
}
}

\seealso{
Other setup helpers: 
\code{\link{use_codecov2}()},
\code{\link{use_vscode}()}
}
\concept{setup helpers}
