% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_muggle_package}
\alias{create_muggle_package}
\title{Create a muggle package}
\usage{
create_muggle_package(
  path,
  fields = list(),
  license = usethis::use_mit_license,
  license_holder = character(),
  organisation = NULL,
  private = FALSE
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}

\item{license}{one of the license functions in \link{usethis}}

\item{license_holder}{giving the license holder, used as \code{cph} and \code{fnd} role in \code{DESCRIPTION}}

\item{organisation}{If supplied, the repo will be created under this
organisation, instead of the account of the user associated with the
\code{auth_token}. You must have permission to create repositories.}

\item{private}{If \code{TRUE}, creates a private repository.}
}
\description{
Set up, or migrate to a muggle project.
Wraps the following steps, \emph{if the respective files or configuration do not already exist}:
\enumerate{
\item \strong{Package Structure}: Sets up scaffolding via \code{\link[usethis:create_package]{usethis::create_package()}} and asks the user to complete the \code{DESCRIPTION}.
\item \strong{Editors/IDEs}: Sets up \href{http://code.visualstudio.com}{vscode} and RStudio as editors.
\item \strong{Git/GitHub}: Initialises a git repo via \code{\link[usethis:use_git]{usethis::use_git()}}, creates a repo on GitHub and sets it as an origin remote.
\item \strong{README}: Adds a \code{README.md} via \code{\link[usethis:use_readme_rmd]{usethis::use_readme_md()}} and asks the user to complete it.
\item \strong{Testing}: Sets up the project for unit tests via \code{\link[usethis:use_testthat]{usethis::use_testthat()}} and test coverage via \code{\link[usethis:use_coverage]{usethis::use_coverage()}}.
\item \strong{Documentation}: Sets up markdown support in roxygen via \code{\link[usethis:use_roxygen_md]{usethis::use_roxygen_md()}}, package documentation via \code{\link[usethis:use_package_doc]{usethis::use_package_doc()}} and ddds a pkgdown website via \code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}}.
\item \strong{Workflow Automation}: sets up caching at \link{lib_cache_path} and tba.
\item \strong{Compute Environment}: tba.
}
}
\section{Warning}{
\itemize{
\item Must not be run \emph{inside} a package, but at the root of all packages
\item If run on an existing project, the project should be under version control, with a clean working tree.
The user should check all changes.
}
}

\concept{setup functions}
