% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_muggle_file}
\alias{get_muggle_file}
\title{Muggle files}
\usage{
get_muggle_file(muggle_file)
}
\arguments{
\item{muggle_file}{File to copy, relative path from built package root.}
}
\description{
Muggle files
}
\keyword{internal}
