% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{get_url_from_desc}
\alias{get_url_from_desc}
\title{Retrieve the public URL from the \code{DESCRIPTION}}
\usage{
get_url_from_desc()
}
\description{
Chooses whichever URL in the \code{DESCRIPTION} is \emph{not} on github.com.
That's assumed to be the public-facing website,
such as a pkgdown website on GitHub pages.
}
\keyword{internal}
