% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_pkgs_image_url}
\alias{gh_pkgs_image_url}
\alias{gh_pkgs_image_name}
\alias{gh_pkgs_image_version}
\title{Get the full docker image url for an image on GitHub Packages}
\usage{
gh_pkgs_image_url(
  image_name = gh_pkgs_image_name(target = "runtime"),
  version = gh_pkgs_image_version(),
  repo_spec = repo_spec()
)

gh_pkgs_image_name(target = c("buildtime", "runtime"))

gh_pkgs_image_version()
}
\arguments{
\item{image_name, version}{Image name and version as strings.
Defaults to muggle convention.}

\item{repo_spec}{GitHub repo specification in this form: \code{owner/repo}.
Users should stick to the default; manual entry is only used for testing.}

\item{target}{Build target for multistage muggle builds.
By convention, for a package \code{foo}, {muggle} would build a \code{foo-buildtime} and \code{foo-runtime} for the \code{buildtime} and \code{runtime} docker multistage build targets, respectively.
A \code{buildtime} target will exist for all {muggle} projects, a \code{runtime} target only for projects with deployed runtimes such as a shiny app}
}
\value{
character string
}
\description{
Helpful to quickly run an image locally or deploy it.
See the \href{https://docs.github.com/en/packages/using-github-packages-with-your-projects-ecosystem/configuring-docker-for-use-with-github-packages}{GitHub Packages for Docker documentation} for details.
Forms a URL of the form \code{docker.pkg.github.com/OWNER/REPOSITORY/IMAGE_NAME:VERSION}
Notice how, different from Docker Hub, docker images on GitHub Packages have an \code{IMAGE_NAME} appended to the familiar \code{OWNER/REPOSITORY} pattern.
These \code{IMAGE_NAME}s are immutable (cannot be changed nor deleted) and must be unique per GitHub repository.
}
\section{Functions}{
\itemize{
\item \code{gh_pkgs_image_name}: Get the docker image name conventionally used in muggle projects

\item \code{gh_pkgs_image_version}: Get the sha of the \emph{latest} \verb{git commit} if on GitHub Actions, or the \emph{head} reference (the branch or tag) otherwise (not recommended for reproducibility).
}}

\concept{GitHub Packages}
