% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{install_deps2}
\alias{install_deps2}
\title{Install package dependencies if needed.}
\usage{
install_deps2()
}
\description{
\code{\link[remotes:install_deps]{remotes::install_deps()}} with muggle defaults
}
\examples{
\dontrun{install_deps(".")}
}
\seealso{
Other computing environment: 
\code{\link{install_sysdeps}()}
}
\concept{computing environment}
