% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_github_actions}
\alias{is_github_actions}
\title{Determine if code is running inside GitHub Actions}
\usage{
is_github_actions()
}
\description{
Looks for the \code{GITHUB_ACTIONS} environment variable, as
\href{https://docs.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables}{documented}
}
\keyword{internal}
