% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_env.R
\docType{data}
\name{lib_cache_path}
\alias{lib_cache_path}
\title{Directory for copying dependencies to docker build context}
\format{
An object of class \code{fs_path} (inherits from \code{character}) of length 1.
}
\usage{
lib_cache_path
}
\description{
This directory serves to copy the package cache into the docker build context on GitHub actions.
}
\seealso{
Other compute environment functions: 
\code{\link{gh_pkgs_image_url}()},
\code{\link{install_sysdeps}()},
\code{\link{use_onbuild_image}()}
}
\concept{compute environment functions}
\keyword{internal}
