% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{local_muggle_file}
\alias{local_muggle_file}
\title{Temporarily get muggle files}
\usage{
local_muggle_file(muggle_file, .local_envir = parent.frame())
}
\arguments{
\item{muggle_file}{File to copy, relative path from built package root.}

\item{.local_envir}{\verb{[environment]}\cr The environment to use for scoping.}
}
\description{
Copies muggle files (\code{\link[=get_muggle_file]{get_muggle_file()}}) to the working directory.
Files are deleted when \code{.local_envir} expires.
If file already exists, only a warning is thrown.
Useful to avoid pasting boilerplate files in muggle packages.
}
\keyword{internal}
