% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{override_pkgdownyaml}
\alias{override_pkgdownyaml}
\title{List of overrides for \verb{_pkgdown.yml} with muggle defaults.}
\usage{
override_pkgdownyaml()
}
\description{
List of overrides for \verb{_pkgdown.yml} with muggle defaults.
}
\examples{
\dontrun{
override_pkgdownyaml()
}

}
\seealso{
Other pkgdown functions: 
\code{\link{build_site2}()},
\code{\link{local_siteyaml}()}
}
\concept{pkgdown functions}
