% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rcmdcheck2}
\alias{rcmdcheck2}
\title{Run R CMD check from R and Capture Results}
\usage{
rcmdcheck2(error_on = "warning")
}
\arguments{
\item{error_on}{Whether to throw an error on \code{R CMD check} failures.
Note that the check is always completed (unless a timeout happens),
and the error is only thrown after completion. If \code{"never"}, then
no errors are thrown. If \code{"error"}, then only \code{ERROR} failures
generate errors. If \code{"warning"}, then \code{WARNING} failures generate
errors as well. If \code{"note"}, then any check failure generated an
error.}
}
\value{
An S3 object (list) with fields \code{errors},
\code{warnings} and \code{notes}. These are all character
vectors containing the output for the failed check.
}
\description{
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} with muggle defaults
}
\concept{helpers for repeated non-interactive use}
