% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{repo_spec}
\alias{repo_spec}
\title{Get the GitHub remote associated with a path as a repo_spec (\code{user/repo})}
\usage{
repo_spec()
}
\description{
Wraps \code{\link[gh:gh_tree_remote]{gh::gh_tree_remote()}}.
}
\concept{GitHub helpers}
\keyword{internal}
