% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{use_codecov2}
\alias{use_codecov2}
\title{Set up codecov}
\usage{
use_codecov2(reposlug)
}
\arguments{
\item{reposlug}{\verb{[character(1)]}
giving the \code{username/repo} URL slug of the project.}
}
\description{
Set up codecov
}
\seealso{
Other quality control functions: 
\code{\link{use_superlinter}()}
}
\concept{quality control functions}
