% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_env.R
\name{use_onbuild_image}
\alias{use_onbuild_image}
\title{Set up or update the compute environment}
\usage{
use_onbuild_image()
}
\description{
Adds or updates a \code{Dockerfile} and corresponding \code{main.yaml} GitHub Actions workflow.
}
\seealso{
Other compute environment functions: 
\code{\link{gh_pkgs_image_url}()},
\code{\link{install_sysdeps}()},
\code{\link{lib_cache_path}}
}
\concept{compute environment functions}
