% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_package2}
\alias{use_package2}
\title{Depend on another package}
\usage{
use_package2(package, type = "Imports")
}
\arguments{
\item{package}{Name of package to depend on.}

\item{type}{Type of dependency: must be one of "Imports", "Depends",
"Suggests", "Enhances", or "LinkingTo" (or unique abbreviation). Matching
is case insensitive.}
}
\description{
\code{\link[usethis:use_package]{usethis::use_package()}} with muggle defaults
}
\examples{
\dontrun{
use_package("ggplot2")
use_package("dplyr", "suggests")
use_dev_package("glue")
}
}
\seealso{
The \href{https://r-pkgs.org/description.html#dependencies}{dependencies section} of \href{https://r-pkgs.org}{R Packages}.
}
\concept{helpers for repeated interactive use}
