% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{use_radian}
\alias{use_radian}
\title{Set up \href{https://github.com/randy3k/radian}{radian}}
\usage{
use_radian()
}
\description{
Set up \href{https://github.com/randy3k/radian}{radian}
}
\concept{editor functions}
