% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{use_superlinter}
\alias{use_superlinter}
\title{Set up \href{https://github.com/github/super-linter}{super-linter}}
\usage{
use_superlinter()
}
\description{
Set up \href{https://github.com/github/super-linter}{super-linter}
}
\seealso{
Other quality control functions: 
\code{\link{use_codecov2}()}
}
\concept{quality control functions}
