% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_vscode}
\alias{use_vscode}
\title{Set up vscode inside a package}
\usage{
use_vscode()
}
\description{
Set up vscode inside a package
}
\details{
\itemize{
\item adds vscode files to \code{.Rbuildignore}
}
}
\seealso{
Other setup helpers: 
\code{\link{create_muggle}()},
\code{\link{use_codecov2}()}
}
\concept{setup helpers}
