% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{view_url}
\alias{view_url}
\title{Browse to URL}
\usage{
view_url(..., open = interactive())
}
\arguments{
\item{...}{Elements of the URL}

\item{open}{\verb{[logical(1)]} giving whether the URL should be opened}
}
\description{
Browse to URL
}
\details{
This function is copied from an unexported function in
\href{https://github.com/r-lib/usethis/blob/23dd62c5e7713ed8ecceae82e6338f795d30ba92/R/helpers.R}{usethis}.
}
\keyword{internal}
